+++ src/led_syscalls.h
@@ -3,7 +3,7 @@
 
 void led_syscalls_init();
 void led_syscalls_frame_start(lssl_vm_t *vm, vm_error_t *error);
-void led_syscalls_calculate_led(lssl_vm_t *vm, int32_t led, float time, vm_error_t *error);
+void led_syscalls_calculate_led(lssl_vm_t *vm, int32_t led, double time, vm_error_t *error);
 void led_syscalls_get_rgb(uint8_t *r, uint8_t *g, uint8_t *b);
 void led_syscalls_get_rgbw(uint8_t *r, uint8_t *g, uint8_t *b, uint8_t *w);
 void led_syscalls_clear();
